+++ websocket-item/src/main/java/com/example/websocketitem/mapper/CommentMapper.java
@@ -1,14 +1,15 @@
 package com.example.websocketitem.mapper;
-import java.util.Date;
+import java.time.LocalDateTime;
 import org.apache.ibatis.annotations.Param;
-import java.util.List;
 
 import com.example.websocketitem.model.Comment;
+
+import java.util.List;
 
 /**
 * @author ASUS
 * @description 针对表【tcd_comment(评论表)】的数据库操作Mapper
-* @createDate 2023-08-25 14:42:48
+* @createDate 2023-08-26 15:23:03
 * @Entity com.example.websocketitem.model.Comment
 */
 public interface CommentMapper {
@@ -27,13 +28,13 @@ public interface CommentMapper {
 
     List<Comment> selectAll();
 
-    List<Comment> selectByArticleIdAndCommentLevelOrderByTopStatusDescAndCreateTimeDesc(@Param("articleId") Integer articleId, @Param("commentLevel") Integer commentLevel);
+    List<Comment> selectByArticleIdAndCommentLevelOrderByTopStatusDescAndCreateTimeDesc(@Param("articleId") Long articleId, @Param("commentLevel") Integer commentLevel);
 
-    List<Comment> selectByParentCommentIdAndArticleIdAndCommentLevelOrderByCreateTimeDesc(@Param("parentCommentId") Integer parentCommentId, @Param("articleId") Integer articleId, @Param("commentLevel") Integer commentLevel);
+    List<Comment> selectByParentCommentIdAndArticleIdAndCommentLevelOrderByCreateTimeDesc(@Param("parentCommentId") Long parentCommentId, @Param("articleId") Long articleId, @Param("commentLevel") Integer commentLevel);
 
-    int updatePraiseNumByCommentId(@Param("praiseNum") Integer praiseNum, @Param("commentId") Integer commentId);
+    int updatePraiseNumByCommentId(@Param("praiseNum") Integer praiseNum, @Param("commentId") Long commentId);
 
-    List<Comment> selectAllByUserId(@Param("userId") Integer userId);
+    List<Comment> selectAllByUserId(@Param("userId") Long userId);
 
     List<Comment> selectByCreateTimeOneWeek();
 }
